+++ Assets/NetRewind/Utils/Simulation/State/StateTypeRegistry.cs
@@ -9,7 +9,7 @@ namespace NetRewind.Utils.Simulation.State
     public static class StateTypeRegistry
     {
         private static bool _initialized;
-        private static readonly Dictionary<Type, int> TypeToId = new();
+        private static readonly Dictionary<Type, ushort> TypeToId = new();
         private static readonly List<Func<IState>> FactoriesById = new();
 
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
@@ -93,17 +93,17 @@ namespace NetRewind.Utils.Simulation.State
             // Deterministic ID assignment
             allMarkedTypes.Sort((a, b) => string.CompareOrdinal(a.FullName, b.FullName));
 
-            for (int i = 0; i < allMarkedTypes.Count; i++)
+            for (ushort i = 0; i < allMarkedTypes.Count; i++)
             {
                 var type = allMarkedTypes[i];
-                int id = i;
+                ushort id = i;
 
                 TypeToId[type] = id;
                 FactoriesById.Add(() => (IState)Activator.CreateInstance(type));
             }
         }
 
-        public static int GetId<T>() where T : IState
+        public static ushort GetId<T>() where T : IState
         {
             InitializeIfNeeded();
 
@@ -114,7 +114,7 @@ namespace NetRewind.Utils.Simulation.State
             return id;
         }
 
-        public static IState Create(int id)
+        public static IState Create(ushort id)
         {
             InitializeIfNeeded();
 
@@ -124,7 +124,7 @@ namespace NetRewind.Utils.Simulation.State
             return FactoriesById[id]();
         }
         
-        public static int GetId(Type type)
+        public static ushort GetId(Type type)
         {
             InitializeIfNeeded();
